Define LD blocks for the SNP panel

Groups SNPs by chromosomal proximity into LD blocks. SNPs on different chromosomes or > 500kb apart are treated as independent. Within a block, a simple uniform correlation structure is used as an approximation (rho = 0.3 between adjacent SNPs, decaying with distance).

Usage

buildLDBlocks(snp_ref)

Value

A list of blocks, each a data.frame of SNPs with their within-block correlation matrix.

Details

In a production version, this would use actual LD matrices from 1000 Genomes EUR via the ieugwasr or LDlinkR package.